Crate deskulpt_logs

This crate implements internal APIs for Deskulpt logs management.

âš ï¸ This crate is meant to be used internally by the Deskulpt application and is not designed to support plugin authors or other users directly. Private items are documented for reference of Deskulpt developers.

Modules

commands ğŸ”’

Tauri commands.

manager ğŸ”’

Deskulpt logs manager and its APIs.

reader ğŸ”’

Logs reading, filtering, and pagination.

Structs

Cursor

Cursor for log pagination.

Entry

A single log entry.

LogsManager

Manager for Deskulpt logs.

Page

A page of log entries.

Traits

LogsExt

Extension to Manager for accessing Deskulpt logs APIs.

Functions

init

Initialize the internal Deskulpt logs plugin.